Project1 WST8/requirements.txt
telebot~=0.0.5
python-dotenv~=1.0.0
reportlab~=4.0.7
gspread~=5.12.0
matplotlib~=3.8.1
oauth2client~=4.1.3
google-api-python-client~=2.107.0
pandas~=2.1.2